+++ modules/multi-runner/variables.tf
@@ -142,7 +142,7 @@ variable "multi_runner_config" {
         group_id                = optional(string)
         group_name              = optional(string)
         host_id                 = optional(string)
-        host_resource_group_arn = optional(number)
+        host_resource_group_arn = optional(string)
         spread_domain           = optional(string)
         tenancy                 = optional(string)
         partition_number        = optional(number)
@@ -152,6 +152,7 @@ variable "multi_runner_config" {
         prefix_log_group = bool
         file_path        = string
         log_stream_name  = string
+        log_class        = optional(string, "STANDARD")
       })), null)
       block_device_mappings = optional(list(object({
         delete_on_termination = optional(bool, true)
@@ -324,6 +325,17 @@ variable "logging_kms_key_id" {
   description = "Specifies the kms key id to encrypt the logs with"
   type        = string
   default     = null
+}
+
+variable "log_class" {
+  description = "The log class of the CloudWatch log groups. Valid values are `STANDARD` or `INFREQUENT_ACCESS`."
+  type        = string
+  default     = "STANDARD"
+
+  validation {
+    condition     = contains(["STANDARD", "INFREQUENT_ACCESS"], var.log_class)
+    error_message = "`log_class` must be either `STANDARD` or `INFREQUENT_ACCESS`."
+  }
 }
 
 variable "lambda_s3_bucket" {
@@ -751,4 +763,10 @@ variable "lambda_event_source_mapping_maximum_batching_window_in_seconds" {
   description = "Maximum amount of time to gather records before invoking the lambda function, in seconds. AWS requires this to be greater than 0 if batch_size is greater than 10. Defaults to 0."
   type        = number
   default     = 0
+}
+
+variable "parameter_store_tags" {
+  description = "Map of tags that will be added to all the SSM Parameter Store parameters created by the Lambda function."
+  type        = map(string)
+  default     = {}
 }