+++ modules/runners/pool.tf
@@ -22,6 +22,7 @@ module "pool" {
       log_level                      = var.log_level
       logging_retention_in_days      = var.logging_retention_in_days
       logging_kms_key_id             = var.logging_kms_key_id
+      log_class                      = var.log_class
       reserved_concurrent_executions = var.pool_lambda_reserved_concurrent_executions
       s3_bucket                      = var.lambda_s3_bucket
       s3_key                         = var.runners_lambda_s3_key
@@ -33,6 +34,7 @@ module "pool" {
       runtime                        = var.lambda_runtime
       timeout                        = var.pool_lambda_timeout
       zip                            = local.lambda_zip
+      parameter_store_tags           = local.parameter_store_tags
     }
     pool                      = var.pool_config
     role_path                 = local.role_path