+++ modules/webhook/eventbridge/dispatcher.tf
@@ -39,7 +39,7 @@ resource "aws_lambda_function" "dispatcher" {
       for k, v in {
         LOG_LEVEL                                = var.config.log_level
         POWERTOOLS_LOGGER_LOG_EVENT              = var.config.log_level == "debug" ? "true" : "false"
-        POWERTOOLS_SERVICE_NAME                  = "dispatcher"
+        POWERTOOLS_SERVICE_NAME                  = "${var.config.prefix}-dispatcher"
         POWERTOOLS_TRACE_ENABLED                 = var.config.tracing_config.mode != null ? true : false
         POWERTOOLS_TRACER_CAPTURE_HTTPS_REQUESTS = var.config.tracing_config.capture_http_requests
         POWERTOOLS_TRACER_CAPTURE_ERROR          = var.config.tracing_config.capture_error
@@ -73,6 +73,7 @@ resource "aws_cloudwatch_log_group" "dispatcher" {
   name              = "/aws/lambda/${aws_lambda_function.dispatcher.function_name}"
   retention_in_days = var.config.logging_retention_in_days
   kms_key_id        = var.config.logging_kms_key_id
+  log_group_class   = var.config.log_class
   tags              = var.config.tags
 }